refactor: rename dynamo_check.py to sanity_check.py
- Rename deploy/dynamo_check.py to deploy/sanity_check.py
- Update README.md to mention sanity check before trying out Dynamo
- Update usage instructions in sanity_check.py to reflect new path
